Fix 4 own C4244/C4267 narrowing-conversion warnings with explicit static_cast
RecursiveLinearFilter.cpp (size_t->long circular buffer indices, 4 sites),
ImpulseResponse.cpp (double->float normalization gain), Resample.h
(double->T cubic interpolation result), wav.cpp (int->uint32_t subFormat).
No behavior change -- narrowed values already fit the target types in every
real case; this only silences the warning with explicit intent.

Deliberately not touched: C4505 warnings in Dependencies/WDL/wdltypes.h --
vendored third-party code, not fork-owned, different authorization bar per
AGENTS.md.

User-confirmed rule this round: fix every warning encountered, not just
ones already flagged by name.

+++ dsp/ImpulseResponse.cpp
@@ -77,7 +77,7 @@ void dsp::ImpulseResponse::_SetWeights()
   // Gain reduction.
   // https://github.com/sdatkinson/NeuralAmpModelerPlugin/issues/100#issuecomment-1455273839
   // Add sample rate-dependence
-  const float gain = pow(10, -18 * 0.05) * 48000 / mSampleRate;
+  const float gain = static_cast<float>(pow(10, -18 * 0.05) * 48000 / mSampleRate);
   for (size_t i = 0, j = irLength - 1; i < irLength; i++, j--)
     this->mWeight[j] = gain * this->mResampled[i];
   this->mHistoryRequired = irLength - 1;
